Move meta types to core
