+++ saleor/graphql/account/types.py
@@ -14,10 +14,10 @@ from ..checkout.types import Checkout
 from ..core.connection import CountableDjangoObjectType
 from ..core.fields import PrefetchingConnectionField
 from ..core.types import CountryDisplay, Image, PermissionDisplay
+from ..core.types_meta import MetaStore
 from ..core.utils import get_node_optimized
 from ..utils import format_permissions_for_display
 from .enums import CustomerEventsEnum
-from .types_meta import MetaStore
 
 
 class AddressInput(graphene.InputObjectType):
